drivers: i2c: sam0: Add DMA driven mode
Add support for a DMA driven mode to the SAM0 SERCOM I2C driver.
This mode allows the I2C interface to operate using DMA to
supply the device instead of an interrupt for every byte.  This
allows most of the I2C transaction to be accomplished without
CPU intervention.

Signed-off-by: Derek Hageman <hageman@inthat.cloud>