soc: sam0: Use DTS clock definitions
This converts the existing drivers and setup to use clock
definitions from the DTS.  This includes both the clock
frequency and the actual selection of the specific GCLK the
driver uses.

At this stage, only the existing configuration is replicated.
So if XOSC32K is enabled in the DTS it is used as the main
clock, reproducing the existing configuration.  Otherwise
OSC8M is used.

Signed-off-by: Derek Hageman <hageman@inthat.cloud>