sam0: clock_control: Implement SAM0 DTS control control
This implements a SAM0 clock_control driver and converts all
existing peripheral drivers to use it.  Additionally the
existing GCLK configuration are converted to use the device
tree to define their sources and output frequencies.  The
previous Kconfig options that controlled the board clocking
are removed and replaced with simple logic using the DTS
definitions.

Signed-off-by: Derek Hageman <hageman@inthat.cloud>