soc: sam0: Disable conditionally supported TC instances
Not all variants of the SAM0 chips support all the TC instances
available in the family.  Some parts of the ASF HAL define the
required parameters regardless, while others do not, which is why
this wasn't caught initially.  So, for now, simply disable the
instances that are conditionally supported and let the overlays
enable them if they know they're using a supported chip.

Signed-off-by: Derek Hageman <hageman@inthat.cloud>